Updated the nginx passenger vhost config to return 503 when serving the maintenance page.
This involved adding a / location and moving the passenger-specific config into it, which shouldn't be a problem. The location{} block is required for nginx to serve maintenance.html as the 503 page; otherwise, it serves its default page. It's something to do with the fact that top-level if{} blocks are evaluated differently to ones within a location{} block.